+++ running/ui/pagerunning.py
@@ -1,36 +1,34 @@
 import tkinter as tk
 from tkinter import ttk
 
+from ui import Page
 from .pageaddrecord import PageAddRecord
 
 
-class PageRunning(tk.Frame):
+class PageRunning(Page):
     
     
-    def __init__(self, root, previousPage):
+    def __init__(self, root, previousPage, rootLabel="SportsPy > Select User > Running"):
         # Set up
-        super().__init__(root, highlightbackground="pink",
-            highlightthickness=5, padx=50, pady=50)
-        self.labelInfoText = "SportsPy > Select User > Running"
+        super().__init__(root, previousPage, rootLabel, highlightbackground="pink", highlightthickness=5, padx=50, pady=50)
         # Dependent pages
-        self.pages = {
-            "PageAddRecord": PageAddRecord(root, self)
-            }
+        self.pages["PageAddRecord"] = PageAddRecord(root, self)
         # Title label
-        labelTitle = ttk.Label(self, text="Running")
-        labelTitle.grid(row=0, column=0, pady=(0,10))
+        self.labels["title"] = ttk.Label(self, text="Running")
+        self.labels["title"].grid(row=0, column=0, pady=(0,10))
         # Add record button
-        buttonAddRecord = ttk.Button(self, text="Add Record",
-            command = lambda: root.change_page(self, self.pages["PageAddRecord"])
-            )
-        buttonAddRecord.grid(row=1, column=0, sticky="WE")
+        self.buttons["add"] = ttk.Button(self, text="Add Record")
+        self.buttons["add"].grid(row=1, column=0, sticky="WE")
+        self.fbuttons["add"] = lambda: self.root.change_page(self, self.pages["PageAddRecord"])
+        self.set_button_command("add")
         # Add database button
-        buttonEditDatabase = ttk.Button(self, text="Edit Database")
-        buttonEditDatabase.grid(row=2, column=0, sticky="WE")
+        self.buttons["edit"] = ttk.Button(self, text="Edit Database")
+        self.buttons["edit"].grid(row=2, column=0, sticky="WE")
         # Visualize button
-        buttonVisualize = ttk.Button(self, text="Visualize Data")
-        buttonVisualize.grid(row=3, column=0, sticky="WE")
+        self.buttons["plot"] = ttk.Button(self, text="Plot Data")
+        self.buttons["plot"].grid(row=3, column=0, sticky="WE")
         # Back button
-        buttonBack = ttk.Button(self, text="Back",
-            command=lambda:root.change_page(self, previousPage))
-        buttonBack.grid(row=4, column=0, sticky="WE")
+        self.buttons["back"] = ttk.Button(self, text="Back")
+        self.buttons["back"].grid(row=4, column=0, sticky="WE")
+        self.fbuttons["back"] = lambda: root.change_page(self, self.previousPage)
+        self.set_button_command("back")